[CUR-1329] Address Copilot review on coverage gaps
- gaps: fix ID/label mismatch in _uncovered_assertions. collect_gaps passes
  assertion node IDs, but the per-assertion fraction map is keyed by short
  label, so every partially/fully covered assertion was reported as a gap.
  Pass assertion nodes; look up by label, return IDs. Rewrite the regression
  test to exercise the real collect_gaps path (previously it called the helper
  with bare labels, masking the bug).
- mcp: correct _covered_labels_for_req docstring — conducted coverage counts
  only at ~1.0, not "non-zero".
- metrics/summary: IntegratesRollup / AssociateIntegration covered counts are
  summed from now-fractional coverage; type them as float and format via
  fmt_assertion_count in the summary integration tables.

+++ src/elspais/graph/metrics.py
@@ -482,9 +482,12 @@ class IntegratesRollup:
     provenance (REQ-d00252-D).
     """
 
-    implemented_covered: int
+    # Covered counts are summed from per-dimension coverage, which is now a
+    # sum of fractional per-assertion values (REQ-d00069-J), so they may be
+    # non-integer. Totals are assertion counts and stay int.
+    implemented_covered: float
     implemented_total: int
-    verified_covered: int
+    verified_covered: float
     verified_total: int
 
     @property
@@ -543,9 +546,11 @@ class AssociateIntegration:
 
     associate: str  # owning associate repo name
     requirement_count: int  # distinct consumer requirements integrating it
-    implemented_covered: int
+    # Covered counts may be fractional (sums of per-assertion coverage,
+    # REQ-d00069-J); totals are assertion counts and stay int.
+    implemented_covered: float
     implemented_total: int
-    verified_covered: int
+    verified_covered: float
     verified_total: int
